chore(listing): remove the legacy antd data-table system
aggrid migration is complete: every routed dataType resolves to a grid
definition (asserted by registry-coverage.test.ts), so both antd fallbacks
were unreachable.

40 files, ~6.4k lines.
