+++ src/app/app/virtual-lab/[virtualLabId]/[projectId]/data/(browse)/browse/entity/[type]/page.tsx
@@ -70,7 +70,6 @@ export default async function Page({
             dataType={dataType}
             mainTableProps={{
               selectionType: 'checkbox',
-              searchOpenOnMount: true,
             }}
             allowDownload
             allowDelete
