+++ src/app/app/virtual-lab/[virtualLabId]/[projectId]/data/(browse)/browse/entity/circuit/page.tsx
@@ -1,16 +1,12 @@
 import { ExtendedEntitiesTypeDict } from '@/api/entitycore/types/extended-entity-type';
 import { WorkspaceSection } from '@/constants';
-import { BrowseCircuit } from '@/ui/segments/explore/circuit';
+import { BrowseEntityScope } from '@/features/views/listing/browse-entity';
 
 export default async function Page() {
   return (
-    <BrowseCircuit
+    <BrowseEntityScope
       section={WorkspaceSection.Data}
       dataType={ExtendedEntitiesTypeDict.Circuit}
-      mainTableProps={{
-        selectionType: undefined,
-        searchOpenOnMount: true,
-      }}
     />
   );
 }
