+++ src/features/views/listing/browse-entity.tsx
@@ -3,14 +3,13 @@
 import dynamic from 'next/dynamic';
 
 import { getEntityGridDefinition } from '@/features/data-grid/bindings/entitycore';
-
-import { BrowseEntityScopeLegacy } from './browse-entity-legacy';
+import { log } from '@/utils/logger';
 
 import type { ReactElement } from 'react';
 import type { IBrowseEntityGridProps } from '@/features/data-grid/host/browse-entity-grid';
-import type { BrowseEntityScopeProps } from './browse-entity-legacy';
+import type { BrowseEntityScopeProps } from './types';
 
-// AG Grid is client-only, and lazy so neither path pays for the other's bundle.
+// AG Grid is client-only, so the listing is loaded lazily.
 const BrowseEntityGrid = dynamic(
   () => import('@/features/data-grid/host/browse-entity-grid').then((m) => m.BrowseEntityGrid),
   { ssr: false }
@@ -19,13 +18,16 @@ const BrowseEntityGrid = dynamic(
 export type { BrowseEntityScopeProps };
 
 /**
- * Per-entity router: a dataType registered in the data-grid entitycore registry renders
- * the AG Grid listing, everything else the legacy antd table.
+ * Entity listing. Every routed dataType must be registered in the data-grid entitycore
+ * registry — see `registry-coverage.test.ts`, which asserts exactly that for each real
+ * call site. An unregistered type renders nothing rather than silently falling back to
+ * a second implementation.
  */
 export function BrowseEntityScope(props: BrowseEntityScopeProps) {
   const definition = getEntityGridDefinition(props.dataType);
-  if (definition) {
-    return <BrowseEntityGrid {...props} definition={definition} />;
+  if (!definition) {
+    log('error', `[BrowseEntityScope] no grid definition registered for "${props.dataType}"`);
+    return null;
   }
-  return <BrowseEntityScopeLegacy {...props} />;
+  return <BrowseEntityGrid {...props} definition={definition} />;
 }
