--- src/ui/segments/data-table/elements/listing-filter-panel/clear-filters.tsx
@@ -1,16 +1,0 @@
-import ReloadIcon from '@/components/icons/Reload';
-
-export function ClearFilters({ onClick, disabled }: { onClick: () => void; disabled: boolean }) {
-  return (
-    <button
-      type="button"
-      className="flex min-w-fit items-center justify-center gap-2 border-none px-4 py-3 text-white hover:bg-[#ffffff33]"
-      onClick={onClick}
-    >
-      <div>Clear filters</div>
-      <ReloadIcon className="text-white" />
-    </button>
-  );
-}
-
-export default ClearFilters;
