--- src/ui/segments/data-table/elements/listing-filter-panel/filter-as-dropdown.tsx
@@ -1,135 +1,0 @@
-import { Select } from 'antd';
-import { map } from 'es-toolkit/compat';
-import { useCallback, useEffect, useState } from 'react';
-
-import { cn } from '@/utils/css-class';
-
-import type { DefaultOptionType } from 'antd/es/select';
-import type { TCoreFilter } from '@/entity-configuration/definitions/types';
-
-export type OptionType = DefaultOptionType & {
-  description?: string;
-};
-
-type Props = {
-  filter: TCoreFilter;
-  data?: Array<OptionType>;
-  onChange: (values: string[]) => void;
-  placeholder?: string;
-  allowMultiple?: boolean;
-};
-
-export function DropdownList({
-  filter,
-  data = [],
-  onChange,
-  placeholder = 'Select options...',
-  allowMultiple = false,
-}: Props) {
-  const [selectedValues, setSelectedValues] = useState<Array<string>>([]);
-
-  useEffect(() => {
-    if (filter.value) {
-      if (Array.isArray(filter.value)) {
-        setSelectedValues(filter.value as string[]);
-      } else if (typeof filter.value === 'string') {
-        setSelectedValues([filter.value]);
-      }
-    } else {
-      setSelectedValues([]);
-    }
-  }, [filter.value]);
-
-  const handleChange = useCallback(
-    (value: string | string[]) => {
-      const newValues = Array.isArray(value) ? value : [value];
-      setSelectedValues(newValues);
-      onChange(newValues);
-    },
-    [onChange]
-  );
-
-  const handleClear = useCallback(() => {
-    setSelectedValues([]);
-    onChange([]);
-  }, [onChange]);
-
-  const options = map(data, (item) => ({
-    value: item.value,
-    label: item.count ? `${item.label} (${item.count})` : item.label,
-    key: item.id,
-    description: item.description,
-  }));
-
-  let value: string | string[] | undefined;
-  if (selectedValues.length > 0) {
-    if (allowMultiple) {
-      value = selectedValues;
-    } else {
-      [value] = selectedValues;
-    }
-  } else {
-    value = undefined;
-  }
-
-  return (
-    <div className="flex flex-col gap-2">
-      <Select
-        allowClear
-        mode={allowMultiple ? 'multiple' : undefined}
-        value={value}
-        onChange={handleChange}
-        onClear={handleClear}
-        placeholder={placeholder}
-        optionFilterProp="label"
-        className={cn(
-          'w-full',
-          '[&_.ant-select-selector]:border-neutral-3! [&_.ant-select-selector]:bg-transparent!',
-          '[&_.ant-select-selection-placeholder]:text-primary-4! [&_.ant-select-selection-item]:text-white!',
-          '[&_.ant-select-arrow]:text-white! [&_.ant-select-clear]:text-white!',
-          '[&_.ant-select-selection-item]:bg-primary-7!',
-          'has-[.ant-select-clear]:[&_.ant-select-arrow]:hidden!',
-          '[&_.ant-select-clear]:bg-transparent!',
-          '[&_.ant-select-selection-item-remove]:text-white! [&_.ant-select-selection-item-remove]:ml-2!',
-          '[&_.ant-select-selection-item-remove]:hover:text-primary-4!'
-        )}
-        classNames={{
-          popup: {
-            root: cn(
-              'bg-primary-8! border-neutral-3 text-white!',
-              '[&_.ant-select-item]:text-white!',
-              '[&_.ant-select-item-option-selected]:bg-[#0050B3]! [&_.ant-select-item-option-selected]:text-white!',
-              '[&_.ant-select-item-option-active]:bg-[#096DD9]! [&_.ant-select-item-option-active]:text-white! select-none!',
-              '[&_.ant-select-item-option-state]:text-white!',
-              '[&_.ant-select-item-option-active_*]:text-white!',
-              '[&_.ant-select-item-option-selected_*]:text-white!',
-              '[&_.ant-select-item-option-content]:w-full'
-            ),
-          },
-        }}
-        size="large"
-        options={options}
-        optionRender={(option) => {
-          const description = option?.data?.description;
-          if (!description) {
-            return <span className="text-white">{option.label}</span>;
-          }
-
-          return (
-            <div className="group flex w-full flex-col gap-0">
-              <span className="text-white text-base font-semibold">{option.label}</span>
-              <p className="text-primary-3 text-sm whitespace-normal wrap-break-word group-hover:text-primary-8">
-                {description}
-              </p>
-            </div>
-          );
-        }}
-      />
-      {allowMultiple && selectedValues.length > 0 && (
-        <div className="text-primary-3! text-xs">
-          {selectedValues.length} item{selectedValues.length !== 1 ? 's' : ''} selected
-        </div>
-      )}
-    </div>
-  );
-}
