--- src/ui/segments/data-table/elements/listing-filter-panel/filters.module.css
@@ -1,9 +1,0 @@
-.accordionTrigger .chevron {
-  transition-property: transform;
-  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-  transition-duration: 150ms;
-}
-
-.accordionTrigger[data-state="open"] .chevron {
-  transform: rotate(90deg);
-}
