--- src/ui/segments/data-table/elements/listing-filter-panel/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './hooks';
