--- src/ui/segments/data-table/table.tsx
@@ -1,349 +1,0 @@
-'use client';
-
-import { ConfigProvider, Table, type TableProps } from 'antd';
-import { isString } from 'es-toolkit/compat';
-import { useCallback, useLayoutEffect, useRef, useState } from 'react';
-
-import useResizeObserver from '@/hooks/useResizeObserver';
-import useScrollComplete from '@/hooks/useScrollComplete';
-import TableControls from '@/ui/segments/data-table/elements/controls';
-import { useOnCellRouteHandler } from '@/ui/segments/data-table/elements/hooks';
-import {
-  type RenderButtonProps,
-  type TableRowSelectionProps,
-  useRowSelection,
-} from '@/ui/segments/data-table/elements/use-row-selection';
-import {
-  type UseExpandableTableOptions,
-  useExpandableTable,
-} from '@/ui/segments/data-table/expandable-row/use-expandable-table';
-import { classNames } from '@/util/utils';
-import { cn } from '@/utils/css-class';
-
-import type { TableRef } from 'antd/es/table';
-import type { ExpandableConfig, RowSelectionType } from 'antd/es/table/interface';
-import type { ComponentProps, CSSProperties, ReactNode } from 'react';
-import type { TExtendedEntitiesTypeDict } from '@/api/entitycore/types/extended-entity-type';
-import type { EntityCoreIdentifiable } from '@/api/entitycore/types/shared/global';
-import type { WorkspaceContext } from '@/types/common';
-
-import styles from '@/ui/segments/data-table/elements/table.module.css';
-
-export type OnCellClick<T> = (basePath: string, record: T, type: TExtendedEntitiesTypeDict) => void;
-
-export function CustomTH({
-  children,
-  style,
-  onClick,
-  handleResizing,
-  columnWidth,
-  ...props
-}: {
-  children: ReactNode;
-  style: CSSProperties;
-  onClick: () => void;
-  handleResizing: (e: React.MouseEvent<HTMLElement>) => void;
-  columnWidth?: number;
-}) {
-  const modifiedStyle: CSSProperties = {
-    ...style,
-    fontWeight: '500',
-    color: '#434343',
-    verticalAlign: 'baseline',
-    boxSizing: 'border-box',
-    backgroundColor: 'white',
-    ...(columnWidth ? { width: columnWidth, minWidth: columnWidth } : {}),
-  };
-
-  const handleResize = handleResizing
-    ? (e: React.MouseEvent<HTMLElement>) => {
-        e.stopPropagation();
-        e.preventDefault();
-        handleResizing(e);
-      }
-    : undefined;
-
-  return handleResize ? (
-    <th
-      {...props} /* eslint-disable-line react/jsx-props-no-spreading */
-      style={{ ...modifiedStyle, padding: '16px 16px 16px 0px', position: 'relative' }}
-      className="before:content-none!"
-      data-testid="column-header"
-    >
-      <button
-        className={classNames(
-          'inline-flex w-full flex-col items-start',
-          '[&>.ant-table-column-sorters]:inline-flex [&>.ant-table-column-sorters]:flex-none [&>.ant-table-column-sorters]:items-start! [&>.ant-table-column-sorters]:gap-2'
-        )}
-        onClick={onClick}
-        type="button"
-      >
-        {children}
-      </button>
-      <button type="button" className={styles.dragHandle} onMouseDown={handleResize} />
-    </th>
-  ) : (
-    <th
-      {...props} /* eslint-disable-line react/jsx-props-no-spreading */
-      data-testid="column-header"
-      style={modifiedStyle}
-    >
-      {children}
-    </th>
-  );
-}
-
-export function CustomCell({
-  children,
-  style,
-  ...props
-}: {
-  children: ReactNode;
-  style: CSSProperties;
-}) {
-  const modifiedStyle = {
-    ...style,
-    padding: '14px 6pX',
-  };
-
-  return (
-    <td {...props} /* eslint-disable-line react/jsx-props-no-spreading */ style={modifiedStyle}>
-      {children}
-    </td>
-  );
-}
-
-type AdditionalTableProps<T> = {
-  hasError?: boolean;
-  onCellClick?: OnCellClick<T>;
-};
-
-export function BaseTable<T extends EntityCoreIdentifiable>({
-  columns,
-  dataSource,
-  hasError,
-  loading,
-  onCellClick,
-  rowSelection,
-  showLoadMore,
-  scrollable = true,
-  sticky,
-  expandableConfig,
-  rowClassName,
-  tableStyle,
-  className,
-  onRow,
-  dataType,
-  wrapperClassname,
-  pagination = false,
-  size,
-  id,
-  dataTestId,
-}: TableProps<T> &
-  AdditionalTableProps<T> & {
-    id?: string;
-    dataTestId?: string;
-    showLoadMore?: (value?: boolean) => void;
-    scrollable?: boolean;
-    expandableConfig?: ExpandableConfig<T>;
-    tableStyle?: CSSProperties | undefined;
-    dataType: TExtendedEntitiesTypeDict;
-    wrapperClassname?: ComponentProps<'div'>['className'];
-  }) {
-  const [containerDimension, setContainerDimension] = useState<{ height: number; width: number }>({
-    height: 0,
-    width: 0,
-  });
-
-  const tableRef = useRef<TableRef>(null);
-  const wrapperRef = useRef<HTMLDivElement | null>(null);
-  const tableElement: HTMLElement | null | undefined =
-    tableRef.current?.nativeElement.querySelector('.ant-table-body');
-  const headerElement: HTMLElement | null | undefined =
-    tableRef.current?.nativeElement.querySelector('.ant-table-header');
-  const headerHeight = headerElement?.getBoundingClientRect()?.height ?? 0;
-
-  useLayoutEffect(() => {
-    if (!tableRef.current) return;
-    setContainerDimension(tableRef.current.nativeElement.getBoundingClientRect());
-  }, []);
-
-  const onResize = useCallback((target: HTMLElement) => {
-    setContainerDimension(target.getBoundingClientRect());
-  }, []);
-
-  // added new id explore-table-container-for-observable because we are using this component
-  // outside of the explore and we want to resize the table according to the screen size as well
-  useResizeObserver({
-    element: wrapperRef.current ?? undefined,
-    callback: onResize,
-  });
-
-  useScrollComplete({
-    element: tableElement,
-    callback: showLoadMore,
-  });
-
-  const onCellRouteHandler = useOnCellRouteHandler<T>({
-    dataType,
-    onCellClick,
-  });
-
-  // TODO: fix the render component if error found
-  if (hasError) return <div>Something went wrong</div>;
-  if (!columns?.length) return null;
-
-  return (
-    <div
-      id="base-table-wrapper"
-      ref={wrapperRef}
-      className={cn('flex min-h-0 grow flex-col overflow-hidden', wrapperClassname)}
-    >
-      <ConfigProvider theme={{ hashed: false }}>
-        <Table
-          id={id}
-          data-testid={dataTestId}
-          size={size}
-          ref={tableRef}
-          sticky={sticky}
-          style={tableStyle}
-          aria-label="listing-view-table"
-          className={cn(styles.table, 'grow [&_.ant-table-sticky-holder]:shadow-md', className)}
-          columns={columns?.map((col) => ({
-            ...col,
-            ...onCellRouteHandler(col),
-          }))}
-          components={{
-            header: {
-              cell: CustomTH,
-            },
-            body: {
-              cell: CustomCell,
-            },
-          }}
-          dataSource={dataSource}
-          loading={loading}
-          pagination={pagination}
-          rowClassName={(row: T, index: number, indent: number) =>
-            classNames(
-              styles.tableRow,
-              isString(rowClassName) ? rowClassName : rowClassName?.(row, index, indent)
-            )
-          }
-          onRow={onRow}
-          rowKey={(row) => row.id}
-          rowSelection={rowSelection?.type ? rowSelection : undefined}
-          scroll={
-            scrollable
-              ? {
-                  x: 'fit-content',
-                  y: Math.max(containerDimension.height - headerHeight, 0),
-                }
-              : { x: 'fit-content' }
-          }
-          expandable={expandableConfig}
-        />
-      </ConfigProvider>
-    </div>
-  );
-}
-
-export function WrapperTable<T extends EntityCoreIdentifiable>({
-  columns,
-  dataSource,
-  hasError,
-  loading,
-  onCellClick,
-  renderButton,
-  selectionType,
-  onRowsSelected,
-  selectedRows,
-  keepSelectionOnScopeChange,
-  getCheckboxProps,
-  scrollable = true,
-  dataKey,
-  expandableConfig,
-  expandableOptions,
-  showExpandButtons = true,
-  rowClassName,
-  tableStyle,
-  onRow,
-  className,
-  dataType,
-  controls,
-  baseTableWrapperClassname,
-  allowDownload,
-  allowDelete,
-  workspace,
-}: TableProps<T> &
-  AdditionalTableProps<T> & {
-    renderButton?: (props: RenderButtonProps<T>) => ReactNode;
-    selectionType?: RowSelectionType;
-    scrollable?: boolean;
-    dataKey: string;
-    /** When `selectedRows` is omitted, selection falls back to `coreSelectedRowsAtom(dataKey)`. */
-  } & TableRowSelectionProps<T> & {
-    expandableConfig?: ExpandableConfig<T>;
-    expandableOptions?: UseExpandableTableOptions<T, T>;
-    showExpandButtons?: boolean;
-    tableStyle?: CSSProperties | undefined;
-    dataType: TExtendedEntitiesTypeDict;
-    controls?: ReactNode;
-    workspace?: WorkspaceContext;
-    baseTableWrapperClassname?: ComponentProps<'div'>['className'];
-    allowDownload?: boolean;
-    allowDelete?: boolean;
-  }) {
-  const {
-    rowSelection,
-    selectedRows: activeSelectedRows,
-    clearSelectedRows,
-  } = useRowSelection({
-    dataKey,
-    selectionType,
-    onRowsSelected,
-    dataSource,
-    selectedRows,
-    getCheckboxProps,
-  });
-
-  const { expandableConfig: generatedExpandableConfig } = useExpandableTable(expandableOptions);
-
-  const finalExpandableConfig =
-    expandableConfig || (showExpandButtons ? generatedExpandableConfig : undefined);
-
-  return (
-    <>
-      <BaseTable<T>
-        dataType={dataType}
-        columns={columns}
-        dataSource={dataSource}
-        hasError={hasError}
-        loading={loading}
-        onCellClick={onCellClick}
-        rowKey={(row) => row.id}
-        rowSelection={rowSelection}
-        scrollable={scrollable}
-        expandableConfig={finalExpandableConfig}
-        rowClassName={rowClassName}
-        style={tableStyle}
-        className={className}
-        onRow={onRow}
-        wrapperClassname={baseTableWrapperClassname}
-      />
-      <TableControls
-        visible
-        renderButton={renderButton}
-        selectedRows={activeSelectedRows}
-        clearSelectedRows={clearSelectedRows}
-        dataType={dataType}
-        workspace={workspace}
-        allowDownload={allowDownload}
-        allowDelete={allowDelete}
-        keepSelectionOnScopeChange={keepSelectionOnScopeChange}
-      >
-        {controls}
-      </TableControls>
-    </>
-  );
-}
